requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[argon2]
python-multipart
pydantic[email]
python-jose[cryptography]